trading_buddy_backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Aleksandr Shestakov",email = "103767386+avtorgenii@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
